requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
httpx>=0.24.0
loguru>=0.7.0
pdfplumber>=0.9.0
pymupdf>=1.23.0
aiofiles>=23.0.0
python-docx>=0.8.11
ollama>=0.1.0
langchain-ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.7.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bandit>=1.7.5